% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quine.R
\name{quine}
\alias{quine}
\title{A Self-Reproducing Function}
\usage{
quine()
}
\description{
\code{quine()} is a non-empty function whose only output is a copy of its own
source code (as a \link[=is.language]{language} object).
}
